Allow "%"PRI to build in C++11 mode.
This is a misdesign in the standard, the committee is debating whether it's
a defect or not, we'll know the result in ten years or so.  In the meantime,
gcc-4.8 and clang(version?) added a fix so formerly legal code compiles (but
with a warning...).  This fix is not yet in gcc-4.7 nor clang-3.1, and since
failing to build on compilers that happen in the wild is a bad idea, let's
kludge it around with spaces.

More info: http://gcc.gnu.org/bugzilla/show_bug.cgi?id=52538