Don't place band members across walls.
They must now be within LOS_SOLID of the leader; if not, try another
position (up to the limit of 1000).

Better would be to allow followers to be placed around nearby corners
but not across walls.  One possibility would be to use (limited-depth)
pathfinding.  Another would be to require LOS_SOLID to any previous
band member, not necessarily the leader.

minmay_rotating_lava might need to be looked at, as this change
greatly affects the distribution of the hell knight band.