A mostly manual Transifex sync.
It looks like it can't do anything even resembling a merge, instead
overwriting everything with its version.  Thus, what we need to do is:
* git checkout $LAST_TRANSIFEX_SYNC
* make tx-pull
* SYNC=`git rev-parse HEAD`
* git checkout master
* git cherry-pick $SYNC
which seems to be scriptable, and will properly handle conflicts.  Assuming
there was a tx-push right after the last sync, that is (it wants to delete
.pl desc for Excruciating Wounds), and it can't handle renames (like .es
A detected closed secret door -> A detected secret door).