Grant Okawaru piety (and gifts) for quality not quantity.
This is currently problematic on dungeon levels up to around D:8 as you find
loads of popcorn interspersed with few actual challenges (which in turn are
often enough to kill you), which gives hardly any piety.

I'm forcing this for now to give us an incentive to do something to monster
generation, if it won't be fixed we can disable this for a yet another
release.

Note that contrary to what some think, this formula would make little sense
for most other gods, especially that they already give more piety for higher
HD enemies.  Making it almost exclusive for hardest battles can be explained
for Okawaru, other gods have no theme reasons to reward recklessly running
into danger.  It'd need major balancing, too.