Let Sunlight linger for a while, don't track monsters with spotlights.
It was strange that the light would magically follow everyone who was in the
initial spot for a duration -- with five independent spotlights no less.
Sunlight now creates a stationary halo instead.

Evaporating water is no longer instant, but the area potentially evaporated
is bigger, to compensate.  The old cross-shaped area is guaranteed, the rest
of haloed stuff (ie, corners of the 3x3 space) have a 50% chance.

Also, the 1/100 chance to spawn a plant per evaporated square of shallow
water is no more, it was too obscure.