Do the whole of X[] save handling as one transaction.
Actually, not even one as it won't be committed until the next time you go
to a new level.  Subsequent writes of a given level in a transaction
overwrite the old uncommitted version.

This sadly can double the size of saves, if someone views all existing levels
in a game.