Fix amulet/ring auto-id interaction with ARTP_*.
If the _maybe_id_jewel() call is passed an artefact property, then
the presence of an unknown artefact amulet should suppress auto-id
of a relevant ring and vice versa.  However, this was happening
even for non-artefacts.

As a particular example (discovered by ChrisOelmueller): wear an
unidentified ring of hunger along with an unidentified normal amulet,
and no other jewellery.  Upon crossing the next hunger threshold, the
ring should, but would not, be identified.  Now it is identified unless
the amulet is an artefact (and thus might have ARTP_METABOLISM).

It probably makes sense to check for any unknown artefact, not just
those in a jewellery slot.  This is not currently handled.