Make emperor scorpions SIZE_GIANT, allowing them to tear through webs.
It's unusual for a branch ending enemy to be disabled by one of the main
features of that branch. One solution would've been to make emperor
scorpions web-immune. I prefer increasing their size, because it allows them
to tear through webs (seems fitting for a freakishly large and powerful
scorpion).

This does mean they are the size of elephants, so the description text might
need rewriting...