Reduce dummy vault balancers.
Levels without vaults are boring.

This commit increases the chance for primary vaults from 1/9 to 1/6, and for
minivaults from 1/3 to 1/2.

Beware of possible balance issues -- vaults tend to provide far more danger,
xp and loot than the rest.  As 0.11 releases early August (and freezes...
when?), if this proves problematic, please revert.