A desc for IOODs themselves.
The player deserved to be told what that white thingy is (easy to take it for
a summon of sorts), and how to avoid it.  It's especially important to convey
that stepping just next to it is not safe, due to quantized movement, and also
due to the "en passant" rule.