Don't place extra loot/monsters in layout_loops_cross.
After taking a closer look at this, and discussing it in ##crawl-dev, I
believe it is best if this layout does not place extra monsters or loot. The
reasons are as follows:

1. Loot is really something that belongs in vaults, not layouts.
2. For several versions of the layout, the loot is hard to notice (it's most
   obvious when you get small rooms).
3. Monsters wander. The guards in the middle of the level won't really stand
   out compared to normal monsters. It's better to just let the level be
   populated normally.

For the third one, it might be possible to make it work by placing a huge
number of monsters in that area. But even then, you'd probably want doors or
something.