Simplify and tweak some Pan layout weights.
I've changed the standard weight for Pan layouts to be 10 (like it is
everywhere else) rather than 1. The one layout that was weight 6 is now 60.

I've also lowered the weights on a couple: octagon levels will now be a bit
less common than others (it was pointed out to me that they were never
common in the past - it depended on which dungeon level you entered Pan
from).

Also, two of the new layout_loops vaults will be a bit less common in Pan,
since they have a very distinctive apperance and tend to have more corridors
than the typical Pan level.