Allow the Vault:8 rune to be placed in any of the subvaults.
This fixes a long-standing bug that was introduced when "prize quadrants"
were added. This caused the rune to only be placed in prize quadrants, even
though all of the regular quadrants supported rune placement. It also
introduced a huge spoiler issue: players who knew the map layouts could
identify prize quadrants, and thus know where the rune was without needing
to explore the level.

The rune now has a 1/4 chance of being in any of the four quadrants.