Hardcode the loading of {stash,wield,runrest,gearset,trapwalk,autofight}.lua
You are free to not use their functionality, load a private modified
version, etc, leaving the default version dormant -- and harmless.

This removes the need to manually update your config whenever such a lua
file is removed or added.  They are a part of the game itself now, being
written in lua is a hidden implementation detail.

There is one file left: lua/kills.lua, it produces spammy dumps not everyone
wants.  Perhaps it should be always loaded, but disabled unless someone uses
an option?