Let ghosts cast Ozocubu's Refrigeration.
This is one of those spells that can potentially kill the ghost that casts
it (it will cast it if it lacks rC). It's no different from other spells of
this sort (ball lightning, fire storm, etc). Ghosts with this spell are
fairly likely to have rC, at least.

Pan lords are forbidden from casting it, since otherwise they will freeze
the vault they spawn in.