Remove (unused) support for vault welcome messages.
The only use of something like this, in Pan unique levels, is done a
different way to reduce text duplication.

I guess the intent of such level messages was to announce the presence of
special rooms like in Nethack, but that would bring confusion for no gain.
And they are already possible via map epilogues.