Fix exiting Hell, AK starting abyss.
The feature was being replaced with DNGN_EXIT_PORTAL_VAULT.  This is
necessary for the DNGN_ESCAPE_STAIRS_UP of Labyrinths, but was skipping
the special-case code for Hell (you.hell_branch) and GDT_GAME_START
(D:1).  This led to the result being invalid and causing either a crash
or an error message and return to D:1, depending on wizmode.

Instead, only special-case the silly branch (Lab), and let everything
else use its native feature.