Rod of lightning.
[Unfinished: no actual damage yet.]

A single zap uses only two mana, evoking it in subsequent turns will take 5
(or less if the rod is spent), spreading the damage upon the whole arc
between, with a boost if you're not aiming at the same place.  Releasing
the trigger (ie, doing any action other than evoking the rod) will stop
the discharge.