Make is_tiles() available on the C++ side.
In webtiles, the same binary can run both, so this function returns whether
this particular game is controlled from console or tiles (there may be
spectators using either, possibly at the same time).

Declared constexpr in other build modes, so if we ever turn on C++11¹ mode,
they'll be as effective as #ifdefs.  Without it, optimized builds may or may
not eliminate code that depends on it, debug builds won't.

[1]. While there are niceties like sane initialization of arrays, etc,
FreeBSD and MacOS use truly ancient compilers, so C++11 is currently not an
option.