Make Octopode ring replacement prompt behave more consistently.
In particular, pressing Esc no longer generates a "remove ring" menu,
but cancels out of the process altogether, and ? brings up a menu (with
cancellation behaving as expected, and selection of a ring to replace
actually replacing the ring).

A small addition to the prompt has been made to indicate this behaviour
(there was no indication of it before).

Addresses Mantis 0005763.