Randomize runes spawned in ZotDef.
The old order didn't make much sense in the first place anyway.

This helps recover games with a bug that destroys stuff on the level.

Out of the two runes that can't spawn in real games, I allowed elven but
not mossy, as the generic tile is acceptable for the former but not latter.