Tri-state autopickup menu.
Bold indicates that autopickup is being forced for the subtype, non-bold
that it is not (i.e. that the +/- shown comes from options such as
autopickup_exception).

Use comma (all items), Ctrl-D <letter>, or Ctrl-D <category> to reset
items to their default un-forced state.

Furthermore, never set force_autpickup automatically.

There are still some minor glitches.  In particular, doing "Ctrl-D *"
as the first action in the menu makes "," not work properly until
the menu is exited and re-entered.  Furthermore, the help needs to
be updated.

Fixes #5772 (but not already-affected saves).