Let b0rsuk know the basics of using Trog (tutorial 5).
Among 23 wins, he lacks any Be, and his highest Be was only XL 8.  Just in
case this was due to language issues, here's a translation of the relevant
tutorial lesson to Polish.

(Seriously, though: this is a test of translating tutorials.  Looks like
quite a few of hints mode messages happen there as well, got to make them
translateable...)