Don't count Heroism nor Ashenzari when identifying weapons by skill.
While it makes sense for a worshipper of a god of battle and war gear to be
good at knowing one's weapons, using this invocation for id of weapons of
other types makes little sense.  And Ashenzari already gives identification
by a different mechanism, this kind of id would remove the need to wield
something cursed.