Disable the "cyrillic" fake language.
It is mindless transliteration, and the effect it produces is so bad it is
hard to read.  Cyrillicization is supposed to be done phonetically.  There
are quite detailed rules how to do that, but some of those are unclear to
me, including important ones like whether English "a" turns into "ei" or "a",
which is described as "in closed syllable" vs "open syllable".  This is a
fault of English pronunciation, of course, but unless someone tells me how
to tell tell these apart programmatically, it's kind of a blocker.  English
syllabization rules are bizarre...