In webtiles, assert that the locale uses UTF-8.
This should (hopefully) not be strictly necessary anymore, but we want
remote players to be able to use non-ASCII characters in notes, map glyphs,
and so on.  As chrootting is usually involved, this is a likely setup error;
I can think of no reason someone would want an ancient locale on a
network-accessible server.

A recent commit fixed websocket communication (always UTF-8), but accessing
local files (config, char dumps, etc) uses locale encoding.