Bog mummy: add bolt of cold, remove death curse.
Bolt of cold might make this monster into a more credible threat (as opposed
to something that just breaks your potions). Swamp lacks ranged damagers,
and cold is a fairly underused element.

As for the death curses: bog mummies are regular corpses preserved in a bog.
They are not like regular mummies (of the ancient Egyptian variety) where it
makes sense that a curse is placed upon them. Bog mummies are closer to
other undead: just another reanimated corpse.