Fix octopode usable tentacle calculation.
It has been suggested that it could be removed, but since it's there, let's try
to fix it first. Octopodes can constrict:
8 monsters empty handed
6 monsters with a one-handed weapon or a shield
5 monsters with a 1.5-handed weapon
4 monsters with a 2-handed weapon, or a weapon and a shield

It's visible in the @ status message.

It might seem needlessly complicated, but it's probably rarely relevant, and is
actually a good way to explicit how octopodes use their limbs.