Rename bog mummies to bog bodies.
Without the death curses, calling them mummies (and putting them on 'M') is
confusing. "Bog body" is good name for this monster (it's synonymous with
bog mummy). I've put it on 'z' and made it lightgreen. They've also been
removed from one ossuary.

Thanks to |amethyst for suggesting this change.