Vary the amount of shallow water in Swamp by depth.
You'll get a bit less on Swamp:1-2, about the same on Swamp:3-4, and
slightly more on Swamp:5. Deep water stays the same. The effect here is
gradual, but you will notice it if you compare Swamp:1 to Swamp:5.

Terrain-wise, this makes the branch somewhat easier than it was before, at
least in the early levels.