Cut abyss CPU usage greatly.
Memoize the feature computed for a cell in the previous turn, rather
than computing it twice per turn.  This allows us to preserve digging
etc. in the abyss without the doubled CPU usage introduced by
0.10-a0-3040-ga00d525.  On my (somewhat slow) machine, the time for a
1000-turn run of the abyss stress test is reduced from ~85 s to ~55 s.

Currently we recompute the cache (i.e., run the worley routines on each
map cell) when loading a saved game and when the abyss shifts as a
result of hitting the map edge.  The last case could be avoided by shifting
around features in the cached array, but abyss shifts are infrequent
enough that this is probably not necessary.