New option: show_travel_trail.
This highlights the path taken during the last exploration or travel,
and should help make travel_delay=-1 less disorienting. The path is
cleared when the player moves otherwise.

In console, it is indicated with COLFLAG_REVERSE, which I'm not sure
is the best option, but just changing the floor colour didn't seem
noticeable enough.