Fix another autopickup menu bug.
It was kind of tricky to trigger, but could result in the ',' command
failing to really reset all the selections.

To reproduce the bug: make sure all AP selections are at the default.
Now enter the menu ('\'), turn on all missiles ('('), enter and leave
the set-to-default mode (Ctrl-D twice), then reset everything to
default (',').  Missiles will not have been reset correctly.