Show the autoexplore limit in Shoals.
The autoexplore is now limited to cells with a height > -18, which is about the
sand limit at low tide. We say that beyond this point, the bottom changes from
sand to rock and autoexplore won't try to go there, even if they are
traverseable.

This greatly reduce autoexplore time, and even amphibious species will respect
it. It doesn't prevent a merfolk from crossing deep water to explore an island
spotted over the water, but they won't randomly explore the sea.

To illustrate the limit, shallow water is light cyan or cyan, deep water is
light blue or blue.