Increase the chance of minivaults being placed in Vaults.
Was 1/2 per level, is now 1/3 (higher than anywhere else except Swamp).
Minivaults usually look alright in the city layouts, and there are no issues
with the level builder having trouble placing them. I haven't touched
primary_mini_dummy because of this - large vaults tend to look weird in this
branch, and fail to place a lot of the time.