Don't double tops of stacks.
We now track in the stash whether we saw a stack, even if we only saw
the top item.  On update, we add the top item if it differs from the
remembered old items, or if we noticed a change from not-a-stack to a
stack.  Previously, the second condition also triggered if we remembered
only a single item of the stack, which would lead to that item being
added to the stash again.

Fixes #5850.  Stacks in old saves will continue to exhibit duplication
until they are verified by visiting the square.