Improve player LRD targetting.
This borrows some code from the monster implementation of LRD,
particularly the split of LRD beam setup from the actual casting of the
spell.

It shows the expected area of effect of the spell (a la the cloud or
storm spells), and warns the player against inadvisable (e.g. friendly
or unaware) targets.