Fix a crash and LRD bad-targetting messages.
We needed to do the ordinary smite-targetting checks before going on to
set up the fragmentation tracer.  For one thing, if an off-map cell was
on-screen, the targetter would crash when checking that cell.  For
another, aiming at an invisible or blocked cell said "You cannot affect
that" (the same as for an un-fragmentable cell) rather than the usual
"You cannot see that place", etc.