Use correct feature name in stashes.
We were looking up the feature description for the item's (x,y)
position... on the current level.  The problem was introduced by
commit 84529d6, but portals had the same problem before then
(which is why  Ctrl-F trove  never worked).

Now we also store the feature's description in the stash.  For old
saves, we use the default description for that type of feature (ignoring
overrides) until we next update the stash.  Floor and other
uninteresting features are stored as the empty string.

Fixes #5880.