Don't display feature backgrounds in most item menus.
Now most item menus, such as the pick-up and stash menus, use the same
background tile as inventory slots.  This is in part to reduce clutter
and make the item easy to see, but also to avoid showing the wrong
background tile for off-level items.

Only Ctrl-X still displays the feature, for consistency with monsters in
the same menu.

Fixes #2303, #5756.