Show meph/evap splash cells as AFF_MAYBE in targetter.
We represent this with a min_expl_rad of 0; a few changes are necessary
to support this.

If the power is 100 or greater (maximum power for Mephitic Cloud,
impossible for Evaporate), you are guaranteed to hit all nine cells, so
display them as AFF_YES (pink rather than purple) then.