Re-add shedu bands to holy pan and zig monster lists.
Also, since Shadow Creatures can now summon a shedu band, make sure that
a worshipper of an evil god angers both shedu at the same time.

Note that other angerable monsters in bands will have the same issue
(only the band leader will be angered)---and it is not possible to use
the same method to solve that problem as we did here.  Therefore, bands
in random monster lists should for the moment be restricted to
non-angerable creatures (other than shedu).

Fixes #5072.

This reverts commit d48ceeeb5743a9a80217dd4b259ec67c4d43fec3.