Fix info leak about unseen wall when targetting bounces.
With the new targetter, you could see bounces off an unseen and unknown
wall.  The spell must be aimed at a seen wall to see bounces in the
first place, but subsequent bounces could leak info.  For example:

      A            B

    #.. #        #.. #
    #K. .        #*. .
    #.. .        #.* <---
--->#..#.        #*.#.
    #..@.        #.*@.

When targetting the indicated wall in A with lightning bolt, you can
see the beam taking the path in B, which reveals that there is a wall
at the indicated position in B.

This was actually a problem before the new targetter, though even harder
to exploit.  If the kobold in A were friendly or otherwise unsafe to
attack, you would get a stop_attack prompt, because the tracer followed
the whole path.

We fix this by making player tracers never bounce off unseen cells, nor
cells the player believes to be non-bouncy.

There is still a minor leak remaining, which is quite a bit harder to
fix: if the cell marked in B is remembered as a wall but has changed
since the last time it was seen (e.g. by a boring beetle), the player
will see the bolt not bouncing.  Likewise, if the cell is unknown but
actually floor, you can see the beam come out the other side.  Fixing
these would probably require player tracers to use map_knowledge rather
than grd in almost all places; that is likely to introduce other
problems.