Fix inconsistencies in handling of some types of death message.
For example, inner flame monster deaths would show "The sheep explodes!
The sheep is blown up!" if the inner flame source is something other
than the player, but only "The sheep explodes!" if the sheep was inner
flamed by the player.

This uses the same set of conditions for display of all death messages.