Handle other branch pairs in grunt_ashenzari_visionary.
It is possible for old saves to have, say, Spider and Snake.  For that
matter, we might change the Lair roulette pairings in the future.  The
vault now handles such situations.  If you happen to have only one
Lair branch for some reason (don't ask me), you'll get its vision on
both sides.