Revert "Show damage numbers in wizard mode."
This reverts commit 1d2dbd5c7c05c53c837a3da882db6ef26e286475.

As discussed on c-r-d, wizmode shouldn't display internals. I think it's fine
to display "spoilers" (like GDR or spell power). They are hidden only for
clarity but since it's possible to calculate them, it's fine to show them in
wizmode.
Displaying damage numbers kinda defeat the use of wizmode for playing a normal
game without permadeath. And it has so much variance it's not even useful
anyway.