Disable self-LRD for now (#5908).
Some but not all of the code from Grunt's monster LRD implementation
made it in (71fa800).  Some code to allow LRDing players made it in,
but not everything necessary.  As a result, setup_fragmentation_beam()
was crashing when called from the targetter.  Temporarily disable
LRD on players until monster LRD (#5840) is merged.