Disable ghost spawning in ZotDef.
This was induced by the move from D:1 to Zot, and resulted in some
unbalanced, if amusing, deaths before the game proper started.

There's been some talk of having ghosts spawn dynamically as part of
waves; this approach is simpler to implement for the time being. Once
the particularls of dynamic spawning have been worked out this can be
revisited.