Have undead base monster selection assume player is in Dungeon in ZotDef.
This restores the behaviour of undead generation prior to the move of
the level from Dungeon to Zot, and prevent horrendously out-of-depth
zombies such as draconians and dragons from being generated within
moments of starting the game.

Eventually, this model could move to being reflective of whatever the
current wave is, but that will require some more work (and possibly
break save compatibility).