Fix confirm_butcher=never choosing useless corpses first.
If you have a poisonous corpse and a regular one on a stack, it doesn't make
any sense to butcher the top one.

This commit will also prefer clean corpses to contaminated ones, as long as
there is no significant difference in freshness.

Finally, it fixes attempting to butcher a forbidden corpse saying there's
nothing there.