IMB tracer and targetting improvements.
IMB tracers now account for the possible explosion, so that the player
will get appropriate warnings and IMB-wielding monsters will consider
the affect that the spell could have on allies. Thus, this fixes
0005823.

There's also now a dedicated targetter for IMB, employing the same code
to determine what squares can be affected as the tracer, which has been
split out from the relevant section of the beam code.

Finally, on account of the above split, the IMB explosion code itself is
now a bit cleaner.