Do a compile check for the presence of fdatasync().
This is a NIH implementation of a part of autoconf, it is limited as it
would be hard to affect the Makefile from such tests.  On the other hand,
even nasty hacks like grepping for sqlite features are not enough here, as
we're looking for something in system headers.

I think all of configuration in the Makefile should be nuked and rewritten,
so it's just a crutch for 0.11.