In ZotDef, don't let monsters wander (and crash) while off-level.
They already get a chance to catch up (returning to the level paralyzes
you for a while), randomly moving them is pointless.  This fixes a crash,
as the orb position was not updated yet, and couldn't be because you were
not on the level yet and it may be in your inventory.

As a side effect, corpses won't rot, rods recharge, etc.  This is consistent
with the notion of no time passing.