Improvements to IMB tracer and targetter.
The tracer now checks for explosions on possible endpoints (i.e. tiles
between the origin and the target where it can hit a monster), and the
targetter shows explosions on those cells.

The display resulting from this is a little bit messy, but the tracer
is now arguably as paranoid as it can be about hitting unwanted targets.