De-clutter IMB target display when orb could hit more than one target.
The path up to the first target is shown as AFF_YES.

Explosion tracers from the first target are shown as AFF_MAYBE.

The path past the first target is also shown as AFF_MAYBE.

The explosion tracers from targets other than the first are now shown as
AFF_TRACER (changed from AFF_MAYBE).

It's thus easier to determine how likely a given square is to be hit
from the display.