Fix rod tiles (#5818).
Subtype and appearance are independent on rods; that makes the normal
hack inapplicable, since it displays the appearance tile for an item
depending on the subtype icon. This makes it very hard to display both
the appearance tile and the subtype icon for a rod on the floor (since
we can only pass one tile index to the rendering code -- that's why
there's the whole mcache thing for monsters).

This is the cowardly fix: Rods on the floor only get the appearance
tile, rods in the inventory menu and region use the same hack as
weapons, armour and ammo (where the brand icon isn't displayed on the
floor either).

I have another fix which displays the subtype icons on the floor, but
it is extremely ugly. The whole tile picking and rendering code could
really use a big refactoring; I don't want to add even more hackiness.