Sprint VI.
There is a lengthy comment that explains the gameplay, but in short: it's a
single room, and monsters spawn in it in a series of increasingly more
difficult battles. The final battle is meant to be all but impossible, but
you are allowed to escape with the orb far before that.

This commit also adds a lua binding for checking if you have the orb of zot.