Don't trigger "monster_dies" triggerers on polymorph.
These are supposed to be triggered when a monster dies or is otherwise
destroyed (with the exception of banishment). Polymorph shouldn't count for
these purposes.

This could be a problem in vaults which use a monster_dies triggerer, if the
player happened to cast polymorph other. It also caused a strange effect with
shapeshifters (they were "dying" every time they transformed).