Sprint VI: Rewrite the way monsters are picked, and increase spawn rate.
Instead of using one huge table of monster specs, there are now several
different ones: for round 1-2, 3-5, 6-10, 11-15, 15-20, 20-30, and 30-57.
This should make it easier to balance things, and make it less likely you'll
get something boring in later rounds.

As for the spawn rate, it now increases rapidly if a round takes more than
100 turns.