Move the shipped font to contribs, upgrade to DejaVu.
Sadly, this fails to handle CJK well; I played long with fontforge trying to
get at least Korean in, but without much success.  In theory, hangul glyphs
are built from few jamo (24, around a hundred graphical variants), but
fontforge seems to copy all 11184 glyphs as combined splines.

It may be possible to have a completely separate font, and load it if the
.ko or .zh translations are selected, but since 99.9% of the game is still
in English (you read descs once), it would need to be a good monospaced
font, and I can't seem to find one.  They're always either proportional or
abysmally bad.

This is an issue only for 0.11, I plan to look into fontconfig (/ pango?)
integration later so the single font requirement or the inability to use
system fonts will be no more.

Also, the way font detection in the Makefile is done is a good example of
something that shouldn't be done in make...