Autoexplore doesn't try to see unreachable places in Shoals.
Autoexplore looping through the whole level was caused by 8d595aab. This
commit prevent this behaviour in Shoals. This is a better and simpler fix
than the autoexplore limit I experimented. It was a bit more efficient, but
less exhaustive, and a bit confusing.

Now that I think about it, we could use a similar fix for swamp instead of
making mangroves fully opaque. Visually, seeing only one row makes it look
less like a forest.