Beogh monster god abilities (read: smiting).
Yes, this means that higher-level orcs (knights, warlords, and
sorcerers) now get smite attacks.

As such, The monster smiting formula's been tweaked to depend on the
monster's "invocations skill", which is the monster's HD for priests and
HD/3 for non-priests. It works out about the same for orc priests (11
damage on average vs the old 12-13), whereas tougher monsters will do
more smiting damage (especially notable with Saint Roka). It's pretty
wimpy for non-priests (sorcerers and knights do about the same smite
damage as a plain priest).