Don't treat "Skybugg" and "Foobola" as buggy.
The former occurred as a randart name in a game ( !lg kryft nawz 2 ).
I suspect "bola" can also occur in random names; it is no longer used
for illegal fixedarts, so there is no need to check for it at all.

Theoretically "buggy" and "buggi" could still occur in random names;
if so, they'll still be reported by the item scan.