Don't use objects with strange constructors for storing monster resists.
This is the same concept as in neunon's branch "smaller-binary", except it
goes farther, getting rid of a list of objects with transposed resist types
and levels.  The old code had an array like:
"+++: rC, rF.  ++:rElec.  -:rPois", which existed only in static constructors
which then were converted again to a regular struct.

It turns out the whole exercise is nearly pointless: savings are around 27K
rather than neunon's 674K; apparently he uses an ANCIENT compiler on a Mac
that can't handle static object initializers well.