Make Ashenzari's lair branch visions less spoilery.
Explicitly state what you are seeing a vision of: "the Shoals and the
Snake Pit" for example.  Also take a note about the situation.

We hard-code the branch names here; better and more translatable would
be to get the long name from branches[], but that doesn't appear to be
available from Lua currently.