Add Iskenderun's Arcane Familiar spell
This is a level 4 conjurations/charms spell which conjures an arcane familiar that
follows the caster around and fires blasts of energy at its creator's target whenever
they cast any projectile-based conjuration.

Like orbs of destruction, they cannot be directly commanded, but should generally
do their job with minimal fuss on the player's part. Unlike orbs of destruction,
they can be targetted and destroyed by enemies.

The familiar is created with a small number of charges. Each time it fires, a charge
is consumed, and the familiar will expire when all these charges are expended
(or after a moderate length of time, if they are not). Only one familiar can be
active at a time, but recasting the spell while one exists will give the familiar
additional charges (up to a limit) and also recall it if out of sight.

If it does not have a clear shot at the chosen target, it will attempt to reposition
itself so that it does; if it can reach a valid spot before the next player turn, it
will fire its own volley, otherwise no shot will be made (though it will hopefully
be better positioned so that it can mirror your spell on the following turn).