Lua function: dgn.br_entrance.
This returns a level ID signifying where the entrance to a branch is
(if not specified, it uses the current branch).

Its primary use would be in the context,

"you.where() == dgn.level_name(dgn.br_entrance("<branch>"))"

to determine if a branch entrance needs to be placed in an encompass
vault or a layout.