Allow encompass vaults to designate locations for Abyss/Hell/Pan portals.
The entry portals for these branches now have a new tag,
"replace_portal"; this instructs the builder to look for Lua markers
with the "portal" property and, if it finds any, to place the portal at
that location (overwriting anything else present).

Best used when stone arches are at the designated locations.