Beef up grunt_snake_rune_spirals somewhat.
This places, on average, an extra 4 nagas, 4 naga warriors, 4 greater
nagas, and 4 guardian serpents; the overall XP total is now roughly on
par with snake_hunt.

This is motivated by players often only clearing half of the ending, in
which case it's much less challenging than it was intended to be;
clearing half the map is now roughly on par with snake_pit.

See https://crawl.develz.org/tavern/viewtopic.php?f=8&t=5217 for more
analysis.