Avoid crash when viewing morgue file with long lines
We attempt to strip the newline character from the end of each line of
the morgue file, but lines longer than 200 characters do not fit into
the input buffer, so there is no newline to strip.

Fix the issue by checking for this condition and omitting the line
truncation when appropriate.