Word-wrap the monster list in character dumps.
For full-game characters, having a half page of monsters in view isn't
uncommon, having to scroll horizontally that far is not fun.  Sadly, we
don't know how wide the browser window will be so I'm wrapping at
oh-so-narrow 80 columns.

Also, the wrapper tends to leave dangling articles at the end of a line,
this isn't looking good either.