Don't clear wall/floor flavour when builder replaces a feature.
It's just the feature that's being replaced, so it makes sense to clear
a special feature (or colour) setting, but not walls and floor when,
say, stairs are being fixed up or replaced with another feature.  This
was causing stairs in vaults to have incorrect floor tiles (#6000),
which is no longer the case.