Revert "Don't clear wall/floor flavour when builder replaces a feature."
Stairs/etc having a generic floor in the background might be somewhat ugly,
but it's nothing compared to stairs/shops/etc looking like walls, floor
looking like water, etc.

Let's think about a fix that handles both for trunk, but in stable, this is
just not a good idea unless we're certain it works.

This reverts commit f06427a7378c94925dfd0ddc0dd742764f4c36e2.