Actually fix loading dumps on most architectures.
The only popular one where this works is amd64, and even that reached 50%
use in Linux on desktops/laptops only recently.  Linux i386, Windows, arm*,
etc, are still 32 bit, so are many less popular platforms.

Nearly any use of "long" in portable code is a bug, except for some things
where using a full machine word improves efficiency but you can still handle
shorter words properly.  Elsewhere else, either an int is enough (and you'd
waste memory on 64 bit), or int is not enough and it will overflow on 32 bit.