Let force_autopickup work on bad butchered chunks.
turn_corpse_into_chunks() set ISFLAG_DROPPED for bad chunks; avoid this
when chunks are set to force-autopickup.

Really it would make sense to allow autopickup_exceptions to override
this as well, but that would probably require doing something like
marking poisonous chunks as bad_item or useless_item (when no relevant
spells are known and the player is not rP).