Nerf Mephitic Cloud :)
("Nerf" as in "make it not crash the game"; the spell itself is
unchanged.)

One argument to targetter_beam::targetter_beam() was missing.
Unfortunately, the subsequent arguments happened to be of either the
correct type or implicitly converted types, and the last two arguments
were optional, so this wasn't caught.  The same was also true for bolt
of cold, but that one didn't cause a crash.

Besides fixing the constructor calls, also update the prototype by
making the optional arguments mandatory; there was only one call that
did not pass those two arguments, anyway.  This should prevent similar
problems in the future, as they will be caught by the compiler.